backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
boto3==1.34.25
python-magic==0.4.27
PyPDF2==3.0.1
ebooklib==0.18
Pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0
